+++ airbyte-server/src/main/java/io/airbyte/server/ServerFactory.java
@@ -13,9 +13,10 @@ import io.airbyte.config.persistence.ConfigPersistence;
 import io.airbyte.config.persistence.ConfigRepository;
 import io.airbyte.db.Database;
 import io.airbyte.scheduler.client.SchedulerJobClient;
-import io.airbyte.scheduler.client.SpecCachingSynchronousSchedulerClient;
+import io.airbyte.scheduler.client.SynchronousSchedulerClient;
 import io.airbyte.scheduler.persistence.JobPersistence;
 import io.airbyte.server.apis.ConfigurationApi;
+import io.airbyte.workers.WorkerConfigs;
 import io.temporal.serviceclient.WorkflowServiceStubs;
 import java.net.http.HttpClient;
 import java.nio.file.Path;
@@ -26,7 +27,7 @@ import org.slf4j.MDC;
 public interface ServerFactory {
 
   ServerRunnable create(SchedulerJobClient schedulerJobClient,
-                        SpecCachingSynchronousSchedulerClient cachingSchedulerClient,
+                        SynchronousSchedulerClient cachingSchedulerClient,
                         WorkflowServiceStubs temporalService,
                         ConfigRepository configRepository,
                         JobPersistence jobPersistence,
@@ -36,6 +37,7 @@ public interface ServerFactory {
                         TrackingClient trackingClient,
                         WorkerEnvironment workerEnvironment,
                         LogConfigs logConfigs,
+                        WorkerConfigs workerConfigs,
                         String webappUrl,
                         AirbyteVersion airbyteVersion,
                         Path workspaceRoot,
@@ -45,7 +47,7 @@ public interface ServerFactory {
 
     @Override
     public ServerRunnable create(final SchedulerJobClient schedulerJobClient,
-                                 final SpecCachingSynchronousSchedulerClient cachingSchedulerClient,
+                                 final SynchronousSchedulerClient synchronousSchedulerClient,
                                  final WorkflowServiceStubs temporalService,
                                  final ConfigRepository configRepository,
                                  final JobPersistence jobPersistence,
@@ -55,6 +57,7 @@ public interface ServerFactory {
                                  final TrackingClient trackingClient,
                                  final WorkerEnvironment workerEnvironment,
                                  final LogConfigs logConfigs,
+                                 final WorkerConfigs workerConfigs,
                                  final String webappUrl,
                                  final AirbyteVersion airbyteVersion,
                                  final Path workspaceRoot,
@@ -66,7 +69,7 @@ public interface ServerFactory {
           jobPersistence,
           seed,
           schedulerJobClient,
-          cachingSchedulerClient,
+          synchronousSchedulerClient,
           new FileTtlManager(10, TimeUnit.MINUTES, 10),
           MDC.getCopyOfContextMap(),
           configsDatabase,
@@ -74,6 +77,7 @@ public interface ServerFactory {
           trackingClient,
           workerEnvironment,
           logConfigs,
+          workerConfigs,
           webappUrl,
           airbyteVersion,
           workspaceRoot,