+++ airbyte-webapp/src/views/layout/MainView/MainView.tsx
@@ -1,6 +1,10 @@
 import React from "react";
 import styled from "styled-components";
+
+import { LoadingPage } from "components";
 import SideBar from "views/layout/SideBar";
+import { StartOverErrorView } from "views/common/StartOverErrorView";
+import { ResourceNotFoundErrorBoundary } from "views/common/ResorceNotFoundErrorBoundary";
 
 const MainContainer = styled.div`
   width: 100%;
@@ -8,6 +12,7 @@ const MainContainer = styled.div`
   overflow: hidden;
   display: flex;
   flex-direction: row;
+  min-height: 680px;
 `;
 
 const Content = styled.div`
@@ -19,7 +24,13 @@ const Content = styled.div`
 const MainView: React.FC = (props) => (
   <MainContainer>
     <SideBar />
-    <Content>{props.children}</Content>
+    <Content>
+      <ResourceNotFoundErrorBoundary errorComponent={<StartOverErrorView />}>
+        <React.Suspense fallback={<LoadingPage />}>
+          {props.children}
+        </React.Suspense>
+      </ResourceNotFoundErrorBoundary>
+    </Content>
   </MainContainer>
 );