+++ airbyte-webapp/src/views/layout/SideBar/SideBar.tsx
@@ -2,13 +2,12 @@ import React from "react";
 import styled from "styled-components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faRocket } from "@fortawesome/free-solid-svg-icons";
-import { faSlack } from "@fortawesome/free-brands-svg-icons";
 import { FormattedMessage } from "react-intl";
 import { NavLink } from "react-router-dom";
 
-import { Routes } from "pages/routes";
+import { RoutePaths } from "pages/routes";
 import { useConfig } from "config";
-import useWorkspace from "hooks/services/useWorkspace";
+import { useCurrentWorkspace } from "hooks/services/useWorkspace";
 
 import { Link } from "components";
 import Version from "components/Version";
@@ -19,6 +18,7 @@ import DocsIcon from "./components/DocsIcon";
 import OnboardingIcon from "./components/OnboardingIcon";
 import SettingsIcon from "./components/SettingsIcon";
 import SourceIcon from "./components/SourceIcon";
+import ResourcesPopup from "./components/ResourcesPopup";
 import { NotificationIndicator } from "./NotificationIndicator";
 
 const Bar = styled.nav`
@@ -91,7 +91,7 @@ const HelpIcon = styled(FontAwesomeIcon)`
 
 const SideBar: React.FC = () => {
   const config = useConfig();
-  const { workspace } = useWorkspace();
+  const workspace = useCurrentWorkspace();
 
   return (
     <Bar>
@@ -99,8 +99,8 @@ const SideBar: React.FC = () => {
         <Link
           to={
             workspace.displaySetupWizard
-              ? Routes.Onboarding
-              : Routes.Connections
+              ? RoutePaths.Onboarding
+              : RoutePaths.Connections
           }
         >
           <img src="/simpleLogo.svg" alt="logo" height={33} width={33} />
@@ -108,7 +108,7 @@ const SideBar: React.FC = () => {
         <Menu>
           {workspace.displaySetupWizard ? (
             <li>
-              <MenuItem to={Routes.Onboarding} activeClassName="active">
+              <MenuItem to={RoutePaths.Onboarding}>
                 <OnboardingIcon />
                 <Text>
                   <FormattedMessage id="sidebar.onboarding" />
@@ -117,7 +117,7 @@ const SideBar: React.FC = () => {
             </li>
           ) : null}
           <li>
-            <MenuItem to={Routes.Connections} activeClassName="active">
+            <MenuItem to={RoutePaths.Connections}>
               <ConnectionsIcon />
               <Text>
                 <FormattedMessage id="sidebar.connections" />
@@ -125,7 +125,7 @@ const SideBar: React.FC = () => {
             </MenuItem>
           </li>
           <li>
-            <MenuItem to={Routes.Source} activeClassName="active">
+            <MenuItem to={RoutePaths.Source}>
               <SourceIcon />
               <Text>
                 <FormattedMessage id="sidebar.sources" />
@@ -133,27 +133,10 @@ const SideBar: React.FC = () => {
             </MenuItem>
           </li>
           <li>
-            <MenuItem to={Routes.Destination} activeClassName="active">
+            <MenuItem to={RoutePaths.Destination}>
               <DestinationIcon />
               <Text>
                 <FormattedMessage id="sidebar.destinations" />
-              </Text>
-            </MenuItem>
-          </li>
-          <li>
-            <MenuItem
-              to={`${Routes.Settings}${Routes.Account}`}
-              activeClassName="active"
-              isActive={(_, location) =>
-                location.pathname.startsWith(Routes.Settings)
-              }
-            >
-              <React.Suspense fallback={null}>
-                <NotificationIndicator />
-              </React.Suspense>
-              <SettingsIcon />
-              <Text>
-                <FormattedMessage id="sidebar.settings" />
               </Text>
             </MenuItem>
           </li>
@@ -169,21 +152,39 @@ const SideBar: React.FC = () => {
           </MenuLinkItem>
         </li>
         <li>
-          <MenuLinkItem href={config.ui.slackLink} target="_blank">
-            {/*@ts-ignore slack icon fails here*/}
-            <HelpIcon icon={faSlack} />
-            <Text>
-              <FormattedMessage id="sidebar.slack" />
-            </Text>
-          </MenuLinkItem>
+          <ResourcesPopup
+            options={[
+              { value: "docs" },
+              { value: "slack" },
+              { value: "recipes" },
+            ]}
+          >
+            {({ onOpen }) => (
+              <MenuItem onClick={onOpen} as="div">
+                <DocsIcon />
+                <Text>
+                  <FormattedMessage id="sidebar.resources" />
+                </Text>
+              </MenuItem>
+            )}
+          </ResourcesPopup>
         </li>
+
         <li>
-          <MenuLinkItem href={config.ui.docsLink} target="_blank">
-            <DocsIcon />
+          <MenuItem
+            to={RoutePaths.Settings}
+            // isActive={(_, location) =>
+            //   location.pathname.startsWith(RoutePaths.Settings)
+            // }
+          >
+            <React.Suspense fallback={null}>
+              <NotificationIndicator />
+            </React.Suspense>
+            <SettingsIcon />
             <Text>
-              <FormattedMessage id="sidebar.docs" />
+              <FormattedMessage id="sidebar.settings" />
             </Text>
-          </MenuLinkItem>
+          </MenuItem>
         </li>
         {config.version ? (
           <li>